Fix bug in Placement constructor
When assigning values to m_grid, the col variable would go 1 place too
far and this corrupted lots of things further down the line.

Refactoring to get the row, col directly from the index is cleaner and
*now* bug-free.

Squishing this bug probably took 5+ hours with Valgrind and minimum
examples. I thought it was related to the Cell *'s.