Updated Routing with S. Added Invalid Location to Terminal.
Might want to either add a copy of nets to original graph, or just take the nets out of cells. I feel like the cells should have terminals with a netID, and the Nets should have pairs of terminals. or something.

Need to do V graph which shouldn't be too hard.

Optimized other things.